:recycle: refactor(auth): restructure code locations
