+++ server/API/Features/Account/Login/Response.cs
@@ -1,3 +1,5 @@
+using server.API.Features.Account.JWT;
+
 namespace server.API.Features.Account.Login;
 
 public class Response
